Add SnipLookupPre and SnipLookupPost User autocmds
These run before/after SnipMate determines its own lookups. If Pre finds
lookups, the normal SnipMate ones are not loaded. Post can be used to
add lookups to the normal set.

This is the first User-facing autocommand, and probably won't be the
last. There will likely need to be some tweaking to how they work.